Modify configuration to allow both drivers and API/library to co-exist. Several refactors were needed in order to make this possible
