+++ src/ApplyOfflineMap.cpp
@@ -15,7 +15,6 @@
 ///	</remarks>
 
 #include "Announce.h"
-#include "CommandLine.h"
 #include "Exception.h"
 #include "OfflineMap.h"
 #include "netcdfcpp.h"
@@ -55,9 +54,10 @@ static void ParseVariableList(
 
 ///////////////////////////////////////////////////////////////////////////////
 
-extern "C" int ApplyOfflineMap( std::string strInputData, std::string strInputMap, std::string strVariables, std::string strInputData2, 
-								std::string strInputMap2, std::string strVariables2, std::string strOutputData, std::string strNColName, 
-								bool fOutputDouble, std::string strPreserveVariables, bool fPreserveAll, double dFillValueOverride
+extern "C" 
+int ApplyOfflineMap( std::string strInputData, std::string strInputMap, std::string strVariables, std::string strInputData2, 
+					 std::string strInputMap2, std::string strVariables2, std::string strOutputData, std::string strNColName, 
+					 bool fOutputDouble, std::string strPreserveVariables, bool fPreserveAll, double dFillValueOverride
 ) {
 
 	NcError error(NcError::silent_nonfatal);
@@ -175,81 +175,5 @@ try {
 }
 	return 0;
 }
-
-///////////////////////////////////////////////////////////////////////////////
-
-#ifdef TEMPEST_DRIVER_MODE
-
-int main(int argc, char** argv) {
-
-	// Input data file
-	std::string strInputData;
-
-	// Input map file
-	std::string strInputMap;
-
-	// List of variables
-	std::string strVariables;
-
-	// Input data file (second instance)
-	std::string strInputData2;
-
-	// Input map file (second instance)
-	std::string strInputMap2;
-
-	// List of variables (second instance)
-	std::string strVariables2;
-
-	// Output data file
-	std::string strOutputData;
-
-	// Name of the ncol variable
-	std::string strNColName;
-
-	// Output as double
-	bool fOutputDouble;
-
-	// List of variables to preserve
-	std::string strPreserveVariables;
-
-	// Preserve all non-remapped variables
-	bool fPreserveAll;
-
-	// Fill value override
-	double dFillValueOverride;
-
-	// Parse the command line
-	BeginCommandLine()
-		CommandLineString(strInputData, "in_data", "");
-		CommandLineString(strInputMap, "map", "");
-		CommandLineString(strVariables, "var", "");
-		CommandLineString(strInputData2, "in_data2", "");
-		CommandLineString(strInputMap2, "map2", "");
-		CommandLineString(strVariables2, "var2", "");
-		CommandLineString(strOutputData, "out_data", "");
-		CommandLineString(strNColName, "ncol_name", "ncol");
-		CommandLineBool(fOutputDouble, "out_double");
-		CommandLineString(strPreserveVariables, "preserve", "");
-		CommandLineBool(fPreserveAll, "preserveall");
-		CommandLineDouble(dFillValueOverride, "fillvalue", 0.0);
-
-		ParseCommandLine(argc, argv);
-	EndCommandLine(argv)
-
-	AnnounceBanner();
-
-	// Calculate metadata
-	int err = ApplyOfflineMap ( strInputData, strInputMap, strVariables, strInputData2, 
-								strInputMap2, strVariables2, strOutputData, strNColName, 
-								fOutputDouble, strPreserveVariables, fPreserveAll, dFillValueOverride );
-	if (err) exit(err);
-
-	// Done
-	AnnounceBanner();
-
-	return 0;
-}
-
-#endif
 
 ///////////////////////////////////////////////////////////////////////////////
