+++ src/GenerateLambertConfConicMesh.cpp
@@ -14,7 +14,6 @@
 ///		or implied warranty.
 ///	</remarks>
 
-#include "CommandLine.h"
 #include "GridElements.h"
 #include "Exception.h"
 #include "Announce.h"
@@ -277,73 +276,5 @@ try {
 }
 	return 0;
 }
-
-///////////////////////////////////////////////////////////////////////////////
-
-#ifdef TEMPEST_DRIVER_MODE
-
-int main(int argc, char** argv) {
-
-	// Number of columns in mesh
-	int nNCol;
-
-	// Number of rows in mesh
-	int nNRow;
-
-	// Reference longitude
-	double dLon0;
-
-	// Reference latitude
-	double dLat0;
-
-	// First standard parallel
-	double dLat1;
-
-	// Second standard parallel
-	double dLat2;
-
-	// Meters to bottom-left X position
-	double dXLL;
-
-	// Meters to bottom-left Y position
-	double dYLL;
-
-	// Cell size
-	double dDX;
-
-	// Output filename
-	std::string strOutputFile;
-
-	// Parse the command line
-	BeginCommandLine()
-		CommandLineInt(nNCol, "ncol", 5268);
-		CommandLineInt(nNRow, "nrow", 4823);
-		CommandLineDouble(dLon0, "lon0", -100.0);
-		CommandLineDouble(dLat0, "lat0", 42.5);
-		CommandLineDouble(dLat1, "lat1", 25.0);
-		CommandLineDouble(dLat2, "lat2", 60.0);
-		CommandLineDoubleD(dXLL,  "xll", -2015000.0, "(meters)");
-		CommandLineDoubleD(dYLL,  "yll", 1785000.0, "(meters)");
-		CommandLineDoubleD(dDX,   "dx", 1000.0, "(meters)");
-		CommandLineString(strOutputFile, "file", "outLCCMesh.g");
-
-		ParseCommandLine(argc, argv);
-	EndCommandLine(argv)
-
-	// Announce
-	AnnounceBanner();
-
-	// Calculate metadata
-    Mesh mesh;
-    int err = GenerateLambertConfConicMesh(mesh, nNCol, nNRow, dLon0, dLat0, dLat1, dLat2, dXLL, dYLL, dDX, strOutputFile);
-	if (err) exit(err);
-
-	// Done
-	AnnounceBanner();
-
-	return 0;
-}
-
-#endif
 
 ///////////////////////////////////////////////////////////////////////////////
